Fix: don't load scripture if there are no proofs
For catechism questions with no proofs we still attempted to load verses from the ESV api.
Since there are no verses to query in this case the ESV api would return an error.
Since we have to proofs to load verses for we can early return before attempting to load verses from the ESV api.